+++ src/features/home/index.ts
@@ -1,0 +1,3 @@
+export { useSplashTimer } from './hooks/useSplashTimer';
+export { SplashScreen } from './ui/SplashScreen';
+export { SelectionScreen } from './ui/SelectionScreen';
